requirements.txt
ffmpeg-python>=0.2.0
matplotlib>=3.5.0
typer>=0.9.0
rich>=13.0.0
tqdm>=4.64.0